Added channel_breakout and included it in hbs_breakout
Channel breakout is an enhanced version of consolidation breakout. Instead of having boxes that we skew, we look directly into tight consolidation channels that start with height percentage of 40% and goes down to 15%. Breakout from this channel is channeé_breakout. If it occurs with confluence, it will fire hbs_breakout (still fired also for consoldiation_breakout and confluence).